DBZ-2337 All snapshots using exported mode
