--- debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/snapshot/ExportedSnapshotter.java
@@ -1,79 +1,0 @@
-/*
- * Copyright Debezium Authors.
- *
- * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package io.debezium.connector.postgresql.snapshot;
-
-import java.time.Duration;
-import java.util.Optional;
-import java.util.Set;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.debezium.connector.postgresql.PostgresConnectorConfig;
-import io.debezium.connector.postgresql.spi.OffsetState;
-import io.debezium.connector.postgresql.spi.SlotCreationResult;
-import io.debezium.connector.postgresql.spi.SlotState;
-import io.debezium.connector.postgresql.spi.Snapshotter;
-import io.debezium.relational.TableId;
-
-/**
- * @author Chris Cranford
- */
-public class ExportedSnapshotter implements Snapshotter {
-
-    private final static Logger LOGGER = LoggerFactory.getLogger(ExportedSnapshotter.class);
-    private OffsetState sourceInfo;
-
-    @Override
-    public void init(PostgresConnectorConfig config, OffsetState sourceInfo, SlotState slotState) {
-        this.sourceInfo = sourceInfo;
-    }
-
-    @Override
-    public boolean shouldSnapshot() {
-        if (sourceInfo == null) {
-            LOGGER.info("Taking exported snapshot for new datasource");
-            return true;
-        }
-        else if (sourceInfo.snapshotInEffect()) {
-            LOGGER.info("Found previous incomplete snapshot");
-            return true;
-        }
-        else {
-            LOGGER.info("Previous exported snapshot completed, streaming logical changes from last known position");
-            return false;
-        }
-    }
-
-    @Override
-    public boolean shouldStream() {
-        return true;
-    }
-
-    @Override
-    public boolean exportSnapshot() {
-        return true;
-    }
-
-    @Override
-    public Optional<String> buildSnapshotQuery(TableId tableId) {
-        return Optional.of("select * from " + tableId.toDoubleQuotedString());
-    }
-
-    @Override
-    public Optional<String> snapshotTableLockingStatement(Duration lockTimeout, Set<TableId> tableIds) {
-        return Optional.empty();
-    }
-
-    @Override
-    public String snapshotTransactionIsolationLevelStatement(SlotCreationResult newSlotInfo) {
-        if (newSlotInfo != null) {
-            String snapSet = String.format("SET TRANSACTION SNAPSHOT '%s';", newSlotInfo.snapshotName());
-            return "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ; \n" + snapSet;
-        }
-        return Snapshotter.super.snapshotTransactionIsolationLevelStatement(newSlotInfo);
-    }
-}
