+++ documentation/modules/ROOT/pages/connectors/postgresql.adoc
@@ -126,21 +126,12 @@ For more information about PostgreSQL logical replication security, see the link
 Most PostgreSQL servers are configured to not retain the complete history of the database in the WAL segments. This means that the PostgreSQL connector would be unable to see the entire history of the database by reading only the WAL. Consequently, the first time that the connector starts, it performs an initial _consistent snapshot_ of the database. The default behavior for performing a snapshot consists of the following steps. You can change this behavior by setting the {link-prefix}:{link-postgresql-connector}#postgresql-property-snapshot-mode[`snapshot.mode` connector configuration property] to a value other than `initial`.
 
 . Start a transaction with a link:https://www.postgresql.org/docs/current/static/sql-set-transaction.html[SERIALIZABLE, READ ONLY, DEFERRABLE] isolation level to ensure that subsequent reads in this transaction are against a single consistent version of the data. Any changes to the data due to subsequent `INSERT`, `UPDATE`, and `DELETE` operations by other clients are not visible to this transaction.
-. Obtain an `ACCESS SHARE MODE` lock on each of the tables being tracked to ensure that no structural changes can occur to any of the tables while the snapshot is taking place. These locks do not prevent table `INSERT`, `UPDATE` and `DELETE` operations from taking place during the snapshot.
-+
-_This step is omitted when `snapshot.mode` is set to `exported`, which allows the connector to perform a lock-free snapshot_.
 . Read the current position in the server's transaction log.
 . Scan the database tables and schemas, generate a `READ` event for each row and write that event to the appropriate table-specific Kafka topic.
 . Commit the transaction.
 . Record the successful completion of the snapshot in the connector offsets.
 
 If the connector fails, is rebalanced, or stops after Step 1 begins but before Step 6 completes, upon restart the connector begins a new snapshot. After the connector completes its initial snapshot, the PostgreSQL connector continues streaming from the position that it read in step 3. This ensures that the connector does not miss any updates. If the connector stops again for any reason, upon restart, the connector continues streaming changes from where it previously left off.
-
-[WARNING]
-====
-It is strongly recommended that you configure a PostgreSQL connector to set `snapshot.mode` to `exported`. The `initial`, `initial only` and `always` modes can lose a few events while a connector switches from performing the snapshot to streaming change event records when a database is under heavy load.
-This is a known issue and the affected snapshot modes will be reworked to use `exported` mode internally (link:https://issues.redhat.com/browse/DBZ-2337[DBZ-2337]).
-====
 
 [id="snapshot-mode-settings"]
 .Settings for `snapshot.mode` connector configuration property
@@ -162,7 +153,7 @@ This is a known issue and the affected snapshot modes will be reworked to use `e
 |The connector performs a database snapshot and stops before streaming any change event records. If the connector had started but did not complete a snapshot before stopping, the connector restarts the snapshot process and stops when the snapshot completes.
 
 |`exported`
-|The connector performs a database snapshot based on the point in time when the replication slot was created. This mode is an excellent way to perform a snapshot in a lock-free way.
+|Deprecated, all modes are lockless.
 
 ifdef::community[]
 |`custom`
@@ -220,14 +211,6 @@ public interface Snapshotter {
      */
     default boolean shouldStreamEventsStartingFromSnapshot() {
         return true;
-    }
-
-    /**
-     * @return true if when creating a slot, a snapshot should be exported, which
-     * can be used as an alternative to taking a lock
-     */
-    default boolean exportSnapshot() {
-        return false;
     }
 
     /**
@@ -2717,7 +2700,7 @@ The following _advanced_ configuration properties have defaults that work in mos
  +
 `initial_only` - The connector performs an initial snapshot and then stops, without processing any subsequent changes. +
  +
-`exported` -  The connector performs a snapshot based on the point in time when the replication slot was created. This is an excellent way to perform the snapshot in a lock-free way. +
+`exported` -  deprecated +
  +
 ifdef::community[]
 `custom` - The connector performs a snapshot according to the setting for the `snapshot.custom.class` property, which is a custom implementation of the `io.debezium.connector.postgresql.spi.Snapshotter` interface. +
